Added simulator

+++ ui_elements/switch.py
@@ -67,7 +67,7 @@ class SwitchPrivate(QObject):
             painter.setBrush(QColor("#30C030"))
 
         painter.drawRoundedRect(
-            r.adjusted(margin, margin, -margin, -margin), r.height() / 2, r.height() / 2
+            r.adjusted(int(margin), int(margin), int(-margin), int(-margin)), int(r.height() / 2), int(r.height() / 2)
         )
 
         self.mGradient.setColorAt(0, button.darker(130))
